@@ -2,45 +2,62 @@
   "name": "mayfield-finances",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "server/server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "NODE_PATH=. nodemon app.js",
-    "build": "webpack --progress --hide-modules",
-    "watch": "webpack --watch",
-    "dev": "webpack-dev-server --inline --hot --open",
-    "v-comp": "node scripts/create-vue-component.js"
+    "start": "NODE_PATH=. NODE_ENV='production' nodemon -r node_modules/dotenv/config server.js",
+    "start-dev": "NODE_PATH=. NODE_ENV='development' nodemon -r node_modules/dotenv/config server.js",
+    "build": "webpack --mode production --progress --hide-modules",
+    "watch": "webpack --watch --mode development",
+    "dev": "webpack-dev-server --mode development --inline --hot --open",
+    "new-comp": "node scripts/create-vue-component.js",
+    "lint": "./node_modules/.bin/eslint .",
+    "redis": "~/redis/src/redis-server ~/redis/my_config.conf",
+    "redis-cli": "~/redis/src/redis-cli"
   },
   "author": "",
   "license": "ISC",
-  "babel-preset-stage-0": "^6.24.1",
   "dependencies": {
     "axios": "^0.17.1",
     "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-es2015": "^6.24.1",
-    "bootstrap": "^4.0.0-beta",
+    "babel-loader": "^7.1.4",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-react": "^6.24.1",
+    "bluebird": "^3.5.1",
+    "bootstrap": "^4.1.1",
     "coinbase": "^2.0.6",
+    "cookie-parser": "^1.4.3",
     "cors": "^2.8.4",
     "css-loader": "^0.28.7",
     "dotenv": "^4.0.0",
+    "eslint-plugin-react": "^7.9.1",
     "express": "^4.16.2",
     "font-awesome": "^4.7.0",
-    "install": "^0.10.2",
+    "html-loader": "^0.5.5",
+    "node-fetch": "^2.1.2",
     "node-sass": "^4.7.2",
     "nodemon": "^1.14.3",
     "pg": "^7.4.0",
     "pug": "^2.0.0-rc.4",
-    "pug-loader": "^2.3.0",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "react-router-dom": "^4.3.1",
+    "redis": "^2.8.0",
     "sass": "^1.0.0-beta.4",
-    "sass-loader": "^6.0.6",
+    "sass-loader": "^7.0.3",
     "style-loader": "^0.19.1",
-    "vue": "^2.5.13",
-    "vue-loader": "^13.6.1",
-    "vue-template-compiler": "^2.5.13",
-    "webpack": "^3.10.0"
+    "webpack-node-externals": "^1.7.2"
   },
   "devDependencies": {
-    "webpack-dev-server": "^2.9.7"
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "eslint": "^4.19.1",
+    "eslint-loader": "^2.0.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "html-webpack-plugin": "^3.2.0",
+    "pug-html-loader": "^1.1.5",
+    "webpack": "^4.12.0",
+    "webpack-cli": "^3.0.8",
+    "webpack-dev-server": "^3.1.0"
   }
 }
