@@ -1,6 +1,11 @@
-let client = require('coinbase').Client
-let coinbase = new client({
-  'apiKey': process.env.COINBASE_API, 
+import { Client } from 'coinbase'
+import { promisify } from 'util'
+
+const client = new Client({
+  'apiKey': process.env.COINBASE_API,
   'apiSecret': process.env.COINBASE_SECRET
 })
-module.exports = coinbase
+
+
+client.getAccounts = promisify(client.getAccounts).bind(client)
+export default client
